Function randomize::fill_byte_buffer_from_os_random

pub fn fill_byte_buffer_from_os_random(buf: &mut [u8]) -> Result<(), u32>

Fills a byte buffer with random bytes from an OS-specific source.

This is intended for use in obtaining small seed slices (<=64 bytes) to initialize a PRNG with. Despite that intent, you can attempt to fill a buffer of any size. The function will automatically chunk up the buffer appropriately so that each individual randomness request to the OS doesn’t overwhelm the OS call.

It’s possible that the OS randomness call can fail. In this case, the function will automatically attempt to retry up to 10 times. If failure persists, then this function will give up and pass the Err back to you. When this happens, some unknown amount of the buffer might have been filled with randomness before things broke.

The OS Call Used Varies By Target

Windows: BCryptGenRandom. On failure, the Err holds the GetLastError value.

Unix: getrandom. On failure, the Err value will be u32::MAX.

Other: This function will fail to link if you don’t have either windows or unix configured.

Because cargo does not handle target-conditional features very well, this function always exists as long as the os_random crate feature is enabled. However, if you do actually call this function when building for a target other than windows or unix (including MacOS), then you’ll get a linker error.

In other words, you can safely leave the os_random feature on all the time and still build the crate anywhere, as long as you don’t actually call this function outside of windows or unix.