Struct randomize::PCG32

pub struct PCG32 {
    pub state: u64,
    pub inc: u64,
}

A Permuted Congruential Generator with 32-bit output.

Period: 2**64 when inc is odd, otherwise less

Fields
state: u64

The generator’s state.

This changes with each step of the generator. It’s the generator’s “position” within the output stream.

inc: u64

The generator’s increment.

This doesn’t change as the generator advances. Instead it determines which of the possible output streams the generator will use. Each inc value will give a different ordering of all the possible outputs.

Implementations

impl PCG32

pub const fn new(state: u64, inc: u64) -> Self

Creates a new generator by directly using the value given.

pub fn next_u32(&mut self) -> u32

Generate the next u32 in the sequence.

pub fn jump(&mut self, delta: u64)

Jump the generator the given number of steps forward in the sequence.

This can go x steps forward in only about log2(x) time.

Because the sequence is a loop, you can go “back” by x steps just by passing x.wrapping_neg() to go sufficiently far forward.