Trait randomize::Gen32

pub trait Gen32 {
    fn next_u32(&mut self) -> u32;

    fn next_bool(&mut self) -> bool { ... }

    fn next_u8(&mut self) -> u8 { ... }

    fn next_u16(&mut self) -> u16 { ... }

    fn next_u64(&mut self) -> u64 { ... }

    fn next_f32_unit(&mut self) -> f32 { ... }

    fn next_f32_signed_unit(&mut self) -> f32 { ... }

    fn next_f64_unit(&mut self) -> f64 { ... }

    fn next_f64_signed_unit(&mut self) -> f64 { ... }

    fn next_bounded(&mut self, b: u32) -> u32 { ... }

    fn pick<T>(&mut self, buf: &[T]) -> T
    where
        Self: Sized,
        T: Copy,
    { ... }

    fn pick_ref<'b, T>(&mut self, buf: &'b [T]) -> &'b T
    where
        Self: Sized,
    { ... }

    fn pick_mut<'b, T>(&mut self, buf: &'b mut [T]) -> &'b mut T
    where
        Self: Sized,
    { ... }

    fn shuffle<T>(&mut self, buf: &mut [T])
    where
        Self: Sized,
    { ... }
}

A Generator with 32 bits of output per step.

Required methods

fn next_u32(&mut self) -> u32

Generates the next 32 bits of output.

Loading content...
Provided methods

fn next_bool(&mut self) -> bool

Produce a bool

fn next_u8(&mut self) -> u8

Produce a u8

fn next_u16(&mut self) -> u16

Produce a u16

fn next_u64(&mut self) -> u64

Produce a u64

fn next_f32_unit(&mut self) -> f32

Returns an f32 in the unsigned unit range, [0, 1]

If you’d like [0, 1) then just use this and reroll in the (very unlikely) case that you do get 1.0

fn next_f32_signed_unit(&mut self) -> f32

Returns an f32 in the signed unit range, [-1, 1]

fn next_f64_unit(&mut self) -> f64

Returns an f64 in the unsigned unit range, [0, 1]

If you’d like [0, 1) then just use this and reroll in the (very unlikely) case that you do get 1.0

fn next_f64_signed_unit(&mut self) -> f64

Returns an f64 in the signed unit range, [-1, 1]

fn next_bounded(&mut self, b: u32) -> u32

Gives a value within 0 .. B

This is often more efficient than making a BoundedRandU32 if you don’t need to use a specific bound value more than once.

Panics

If the input is 0.

fn pick<T>(&mut self, buf: &[T]) -> T where
Self: Sized,
T: Copy,

Gets a value out of the slice given (by copy).

The default impl will not pick past index u32::MAX.

fn pick_ref<'b, T>(&mut self, buf: &'b [T]) -> &'b T where
Self: Sized,

Gets a value out of the slice given (by shared ref).

The default impl will not pick past index u32::MAX.

fn pick_mut<'b, T>(&mut self, buf: &'b mut [T]) -> &'b mut T where
Self: Sized,

Gets a value out of the slice given (by unique ref).

The default impl will not pick past index u32::MAX.

fn shuffle<T>(&mut self, buf: &mut [T]) where
Self: Sized,

Shuffles a slice in O(len) time.

The default impl shuffles only the first u32::MAX elements.

Loading content...